requirements.txt
# DHI Transaction Analytics - Production Requirements

# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0

# Database Drivers
asyncpg>=0.29.0
aiosqlite>=0.19.0
neo4j>=5.14.1
redis>=5.0.1
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
python-dotenv>=1.0

# HTTP Clients
httpx>=0.25.2
requests>=2.31.0
aiohttp>=3.9.1

# LLM Integration
openai>=1.3.8
anthropic>=0.7.8

# Data Processing
pandas>=2.1.4
numpy>=1.25.2

# File Processing
aiofiles>=23.2.1
python-multipart>=0.0.6
Pillow>=10.1.0

# Plaid API
plaid-python>=9.1.0

# AI/ML Libraries
sentence-transformers>=2.2
spacy>=3.8
transformers>=4.30.0
torch>=2.0.0

# Audio Processing
openai-whisper>=20231117

# Graph Database
weaviate-client>=4.15

# Monitoring & Security
prometheus-client>=0.19.0
python-jose[cryptography]>=3.3.0

# Production Server
gunicorn>=21.2.0